+++ CHANGELOG.md
@@ -7,13 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Most recent change on the top.
 
 ## Unreleased
-- Update `SoftAdapt` callback to weight loss coefficient updates by the chosen (initial) loss coefficients.
 
 
 ## [0.12.1]
 
 ### Added
 - `MaximumAbsoluteError` metric
+
+### Changed
+- Update `SoftAdapt` callback to weight loss coefficient updates by the chosen (initial) loss coefficients.
 
 ### Fixed
 - `PerTypeScaleShift` model modifier shape bug